Tried to implement f16 tensor core

+++ data/shader/include/mrhe.glsl
@@ -76,7 +76,7 @@ void EncodePosMrhe(const vec3 normPos, const uint mrheInputBaseIndex)
 			{
 				sum += linearLerpFactor[neigh] * neighbourFeatures[(neigh * POS_FEATURE_COUNT) + feature];
 			}
-			renderNeurons[levelBaseIndex + feature] = sum;
+			renderNeurons[levelBaseIndex + feature] = float16_t(sum);
 		}
 	}
 }
